+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3
+2016/10/16
+- Added contents matching
+- Fixed bugs
+
 # v0.2
 2016/10/8
 - Added timestamp and size to list